feat(version-engine): Round 3 — state machine, side-band, layout cleanup
Round 3 of docs/architecture/07-version-engine-todo.md, focused on
audit completeness, Git client UX, and the post-mut-removal package
layout. No engine semantics changed; every existing test still passes.

B6 — version_transactions captures pending + rejected states.
  Pending: the existing _record_pending_conflict path now inserts a
  status='pending_manual_review' version_transactions row before the
  mut_conflicts row, and threads the new transaction_id into the FK
  column so the resolver UI can join cleanly.
  Rejected: cross-scope CrossScopeSubmissionError now also writes a
  status='rejected' version_transactions row with reason=
  'cross_scope_paths_outside_scope', closing the gap where rejected
  submissions disappeared from the ledger.

E1 — receive-pack returns explicit (committed/rejected/pending_resolution)
  outcomes. The ng line is tagged with a 'puppyone-rejected:' /
  'puppyone-pending:' prefix so tooling can disambiguate, and side-band
  channel-2 stderr lines surface 'remote: PuppyOne: ...' hints to the
  git client. The legacy ok= kwarg keeps working for callers that
  haven't migrated. 11 new test_receive_pack_outcomes.py cases.

E2 — pure-Python acceptance test that scoped clones cannot reach
  sibling-scope objects. Catches a real bug discovered in the process:
  ``filtered_commit_tree`` was passing the original ``excludes`` to a
  comparison against scope-prefixed paths, which silently let
  client-configured excludes leak into the scoped view. Tightened the
  docstring on ``is_path_excluded`` so the absolute-path convention
  is no longer ambiguous; the parity matrix tests now align with the
  scoped-clone tests.

C2 — services/ops.py is now adapters/operations/ops_adapter.py.
  Per supplement §4, MutOps is a Protocol Adapter, not a service.
  services/ops.py is a thin re-export shim so the 17 existing call
  sites keep working unchanged.

C3 — graft + project-view-index logic moves out of services/hooks.py
  into application/root_projection.py. Two new public functions
  (rebuild_project_root_after_commit, record_project_view_index_for_commit)
  own the algorithm; services/hooks.py shrinks by ~210 lines and now
  just orchestrates: post-commit-delete cleanup, scope-promote dispatch,
  and websocket fan-out.

C4 — services/direct_writer.py trimmed from 85 → 27 lines. The unused
  apply_mutation() wrapper is gone; only ConcurrentMutationError is
  still re-exported because two routers still import it from the old
  path.

F1 — Connector audit. grep confirms zero mut.* imports, zero
  adapters.mut references, zero publish_scope_update bypasses outside
  the engine. Every connector / ingest write converges on MutOps or
  MutEphemeralClient (which itself routes to the engine after the
  Round 2 source_channel="agent" fix).

Tests: 656 mut_engine + 11 new receive-pack + 4 new scoped-clone
isolation cases pass; same 7 pre-existing baseline failures (config /
sandbox / search / agent-visibility) carry over unchanged.
